Libvec up-to-date

+++ Libvec/src/vector2.c
@@ -6,13 +6,13 @@
 /*   By: wwatkins <wwatkins@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/27 14:23:31 by wwatkins          #+#    #+#             */
-/*   Updated: 2016/01/27 16:27:08 by wwatkins         ###   ########.fr       */
+/*   Updated: 2016/02/09 08:09:02 by wwatkins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libvec.h"
 
-t_vec2	vec2(float x, float y)
+t_vec2	vec2(double x, double y)
 {
 	t_vec2	vec2;
 
@@ -28,7 +28,7 @@ t_vec2i	vec2i(t_vec2 vec2)
 	return (vec2i);
 }
 
-t_vec2	vec2_scale(t_vec2 vec2, float scale)
+t_vec2	vec2_scale(t_vec2 vec2, double scale)
 {
 	vec2_normalize(&vec2);
 	vec2.x *= scale;
@@ -36,14 +36,14 @@ t_vec2	vec2_scale(t_vec2 vec2, float scale)
 	return (vec2);
 }
 
-float	vec2_magnitude(t_vec2 vec2)
+double	vec2_magnitude(t_vec2 vec2)
 {
 	return (sqrt(vec2.x * vec2.x + vec2.y * vec2.y));
 }
 
 void	vec2_normalize(t_vec2 *vec2)
 {
-	float	magnitude;
+	double	magnitude;
 
 	magnitude = vec2_magnitude(*vec2);
 	vec2->x /= magnitude;
